:recycle: #6 Fix import path & Move studyDirection to entities
