+++ backend/src/discipline/entities/disciplineHoured.entity.ts
@@ -1,6 +1,6 @@
 import { Entity, Column, PrimaryGeneratedColumn, ManyToOne } from 'typeorm'
-import { StudyDirection } from '../studyDirection.entity';
 import { Discipline } from './discipline.enitity';
+import { StudyDirection } from './studyDirection.entity';
 
 @Entity({name: 'discipline_houred'})
 export class DisciplineHoured {
