+++ backend/src/discipline/entities/studyDirection.entity.ts
@@ -1,6 +1,6 @@
 import { Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm'
-import { DisciplineHoured } from './entities/disciplineHoured.entity';
-import { Group } from '../group/entities/group.entity';
+import { DisciplineHoured } from './disciplineHoured.entity';
+import { Group } from '../../group/entities/group.entity';
 
 @Entity({name: 'study_direction'})
 export class StudyDirection {
@@ -13,6 +13,4 @@ export class StudyDirection {
 
     @OneToMany(() => DisciplineHoured, houredDisciplines => houredDisciplines.studyDirection)
     houredDisciplines: DisciplineHoured[]
-
-
 }
